Create separate Edit component same as other blocks
